🤘 Move GitHub enterprise query to root cmd
